extract Concerns/tagging trait
