+++ src/Illuminate/Container/Container.php
@@ -8,11 +8,14 @@ use LogicException;
 use ReflectionClass;
 use ReflectionParameter;
 use Illuminate\Support\Arr;
+use Illuminate\Container\Concerns\Tagging;
 use Illuminate\Contracts\Container\BindingResolutionException;
 use Illuminate\Contracts\Container\Container as ContainerContract;
 
 class Container implements ContainerContract
 {
+    use Tagging;
+
     /**
      * The current globally available container (if any).
      *
@@ -68,13 +71,6 @@ class Container implements ContainerContract
      * @var array
      */
     protected $extenders = [];
-
-    /**
-     * All of the registered tags.
-     *
-     * @var array
-     */
-    protected $tags = [];
 
     /**
      * The stack of concretions currently being built.
@@ -428,47 +424,6 @@ class Container implements ContainerContract
                 }
             }
         }
-    }
-
-    /**
-     * Assign a set of tags to a given binding.
-     *
-     * @param  array|string  $abstracts
-     * @param  array|mixed   ...$tags
-     * @return void
-     */
-    public function tag($abstracts, $tags)
-    {
-        $tags = is_array($tags) ? $tags : array_slice(func_get_args(), 1);
-
-        foreach ($tags as $tag) {
-            if (! isset($this->tags[$tag])) {
-                $this->tags[$tag] = [];
-            }
-
-            foreach ((array) $abstracts as $abstract) {
-                $this->tags[$tag][] = $abstract;
-            }
-        }
-    }
-
-    /**
-     * Resolve all of the bindings for a given tag.
-     *
-     * @param  string  $tag
-     * @return iterable
-     */
-    public function tagged($tag)
-    {
-        if (! isset($this->tags[$tag])) {
-            return [];
-        }
-
-        return new RewindableGenerator(function () use ($tag) {
-            foreach ($this->tags[$tag] as $abstract) {
-                yield $this->make($abstract);
-            }
-        }, count($this->tags[$tag]));
     }
 
     /**
